requirements.txt
flask
scikit-learn
nltk
numpy
pypdf
python-docx
pandas
werkzeug
pytesseract
Pillow
PyMuPDF
